Fix setup routine for v1.4.1
Bug introduced in 87b12c37647700f8b29bc866fe93a234af4343e4
Fix included in v1.4.1 release

+++ SoundManager/SchemeMeta.cs
@@ -169,7 +169,8 @@ namespace SoundManager
                     image.Save(SchemeImageFilePath + ".bmp");
                 }
 
-                File.Delete(SchemeImageFilePath);
+                if (File.Exists(SchemeImageFilePath))
+                    File.Delete(SchemeImageFilePath);
 
                 if (File.Exists(SchemeImageFilePath + ".bmp"))
                     File.Move(SchemeImageFilePath + ".bmp", SchemeImageFilePath);
